# swagger-spec.yaml
openapi: 3.0.1
info:
  title: Swagger Passport
  description: This is a sample server for post management
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: anuj@ourdesignz.in
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
...
servers:
  - url: 'http://localhost:3000/api'
    description: Internal staging server for testing
  - url: 'https://localhost:3000/api'
    description: Main (production) server
paths:
...
  /post/create:
    post:
      tags:
        - Post
      summary: To create a Post
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePost'
      responses:
        '201':
          description: Created
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    InternalServerError:
      description: Internal server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnprocessableEntity:
      description: Unprocessable Entity
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
    Unauthorized:
      description: Not authorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
    Errors:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              value:
                type: string
              msg:
                type: string
              param:
                type: string
              location:
                type: string
...
    CreatePost:
      type: object
      properties:
        title:
          type: string
        desc:
          type: string
        files:
          type: array
          items:
            type: string
      required:
        - title
        - description
        - files
...